requirements.txt
telebot==0.0.5
python-decouple==3.8